[BUGFIX] Handle analyzing incomplete IRRE entities
Collapsed IRRE entities do not provide values and relations to be set
in target state. For values that's okay since nothing is changed. For
relations however missing information was interpreted as the removal
of existing relations. In those cases relations are re-used from the
existing source state (if it does exist).